refactor(editor): split BufferLifecycle by responsibility, update callers
Replace the grab-bag BufferLifecycle module with three focused modules:
- BufferRegistry: buffer open/register/track/tab-lookup
- SessionRestore: session persistence and swap file recovery
- Notifications: notification lifecycle and scheduling

Updates all callers to reference the new modules directly instead of
going through defdelegate forwarding stubs in the editor. Moves
maybe_refresh_tool_picker back to the editor as an inline function.
